backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
supabase-pydantic
openai
anthropic
langchain-text-splitters
tenacity
supabase

